+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControlExtensions.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Text;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
+using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.LowLevel;
+using UnityEngine.InputSystem.UI;
 using UnityEngine.InputSystem.Utilities;
 
 ////REVIEW: some of the stuff here is really low-level; should we move it into a separate static class inside of .LowLevel?
@@ -60,7 +62,12 @@ namespace UnityEngine.InputSystem
             if (control == null)
                 throw new ArgumentNullException(nameof(control));
             if (Mathf.Approximately(0, buttonPressPoint))
-                buttonPressPoint = InputSystem.settings.defaultButtonPressPoint;
+            {
+                if (control is ButtonControl button)
+                    buttonPressPoint = button.pressPointOrDefault;
+                else
+                    buttonPressPoint = ButtonControl.s_GlobalDefaultButtonPressPoint;
+            }
             return control.IsActuated(buttonPressPoint);
         }